Use size_t in aufile_set_position

+++ include/rem_aufile.h
@@ -27,4 +27,4 @@ int aufile_write(struct aufile *af, const uint8_t *p, size_t sz);
 size_t aufile_get_size(struct aufile *af);
 size_t aufile_get_length(struct aufile *af, struct aufile_prm *prm);
 int aufile_set_position(struct aufile *af, const struct aufile_prm *prm,
-		off_t pos_ms);
+		size_t pos_ms);
